refactor: move helper functions into a separate file
